Migrate to MSBuild
Migrate from project.json to .csproj.
Things left to do include fixing the ruleset and stylecop.json files not
working, an MSBuild reference conflict in the tests, an integration test
that is no longer working due to a change in BrowserStack, and dotnet
pack not including the source in the symbols package.
Also fixes a warning about the use of "Select" in the Build.ps1 script
in VS Code.